commentCleaning
