+++ public/index.html
@@ -7,35 +7,7 @@
       <meta charset="UTF-8" />
       <meta http-equiv="X-UA-Compatible" content="IE=edge" />
       <meta name="viewport" content="width=device-width, initial-scale=1" />
-      <title>Front End Developer | JavaScript Expert - Freelancer</title>
-      <meta
-        name="description"
-        content="Hire Professional Freelancer &amp; Angular Developer. Building animations and interactive experiences, hi-end solutions, website maintenance &amp; security"
-      />
-      <meta name="robots" content="noodp" />
-      <meta property="og:locale" content="en_US" />
-      <meta property="og:type" content="website" />
-      <meta
-        property="og:title"
-        content="Front End Developer | Angular Expert - Freelancer"
-      />
-      <meta
-        property="og:description"
-        content="Hire Professional Freelancer &amp; Angular Developer. Building animations and interactive experiences, hi-end solutions, website maintenance &amp; security"
-      />
-      <meta
-        property="og:site_name"
-        content="Front End Developer | Angular Expert - Freelancer"
-      />
-      <meta name="twitter:card" content="summary" />
-      <meta
-        name="twitter:description"
-        content="Hire Professional Freelancer &amp; Angular Developer. Building animations and interactive experiences, hi-end solutions, website maintenance &amp; security"
-      />
-      <meta
-        name="twitter:title"
-        content="Front End Developer | Angular Expert - Freelancer"
-      />
+      <title>Fullstack Developers | Team God Dag</title>
       <link rel="stylesheet" href="https://unpkg.com/leaflet@1.7.1/dist/leaflet.css"
       integrity="sha512-xodZBNTC5n17Xt2atTPuE1HxjVMSvLVW9ocqUKLsCC5CXdbqCmblAshOMAS6/keqq/sMZMZ19scR4PsZChSR7A=="
       crossorigin=""/>
@@ -45,8 +17,6 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="backdrop-hook"></div>
-    <div id="drawer-hook"></div>
     <div id="root"></div>
     <!-- emailjs -->
     <script type="text/javascript"
