+++ src/App.scss
@@ -1,14 +1,9 @@
 $primary-color: #ffd700;
 
 @import 'animate.css';
-@import '~loaders.css/src/animations/pacman.scss';
+@import '~loaders.css/src/animations/line-scale-pulse-out-rapid.scss';
 
-@font-face {
-  font-family: 'Helvetica Neue';
-  src: url('./assets/fonts/helvetica-neu.ttf') format('ttf');
-  font-weight: normal;
-  font-style: normal;
-}
+
 
 @font-face {
   font-family: 'La Belle Aurore';
@@ -17,29 +12,6 @@ $primary-color: #ffd700;
   font-weight: normal;
   font-style: normal;
 }
-
-@font-face {
-  font-family: 'Coolvetica';
-  src: url('./assets/fonts/CoolveticaRg-Regular.woff2') format('woff2'),
-    url('./assets/fonts/CoolveticaRg-Regular.woff') format('woff');
-  font-weight: normal;
-  font-style: normal;
-}
-
-@font-face {
-  font-family: 'Roboto';
-  src: url('./assets/fonts/Roboto-Regular.ttf') format('ttf');
-  font-weight: normal;
-  font-style: normal;
-}
-
-@font-face {
-  font-family: 'Indie Flower';
-  src: url('./assets/fonts/IndieFlower-Regular.ttf') format('ttf');
-  font-weight: normal;
-  font-style: normal;
-}
-
 input,
 textarea {
   font-family: 'Helvetica Neue', sans-serif;
