+++ src/components/About/index.js
@@ -1,7 +1,6 @@
 import { useEffect, useState } from 'react'
 import Loader from 'react-loaders'
 import AnimatedLetters from '../AnimatedLetters'
-// import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import './index.scss'
 import EmployeeList from '../EmployeeList'
 
@@ -17,6 +16,7 @@ const About = () => {
 
 
    return (
+      <>
       <div className="container about-page">
          <div className="text-zone">
             <h1>
@@ -28,11 +28,10 @@ const About = () => {
             </h1>
          </div>
 
-         <EmployeeList />
-
-
-         <Loader type="pacman" />
+         <EmployeeList />         
       </div>
+      <Loader type="line-scale-pulse-out-rapid"  />
+      </>
    )
 }
