+++ src/components/About/index.scss
@@ -1,16 +1,68 @@
-// div {
-//     position: relative;
-//     width: 200px;
-//     height: 200px;
-//     border: 1px solid #ccc;
-//     background: rgba(255, 255, 255, 0.4);
-//     text-align: center;
-//     font-size: 100px;
-//     display: flex;
-//     justify-content: center;
-//     align-items: center;
-//     box-shadow: 0 0 20px 0px lightyellow;
-//   }
+.container.about-page {
+  .text-zone {
+     position: relative;
+     left: 66%;
+     top: 0.5rem;
+     width: 100% !important;
+     vertical-align: middle;
+     display: table-cell;
+     max-height: 90%;
+
+     h1 {
+        font-size: 60px;
+        font-family: 'Leckerli One', cursive;
+        color: #FFAE35;
+        font-weight: 400;
+        margin-top: 0;
+        position: relative;
+        left: -10%;
+        top: 90%;
+        transform: translateY(60%);
+        width: 100% !important;
+        vertical-align: middle;
+        display: table-cell;
+        max-height: 90%;
+
+        &:before {
+           content: "<h1>";
+           font-family: "La Belle Aurore", cursive;
+           color: #ffffff;
+           font-size: 1.5rem;
+           position: absolute;
+           margin-top: -10px;
+           left: -10px;
+           opacity: 0.6;
+           line-height: 18px;
+        }
+
+        &:after {
+           content: "<h1/>";
+           font-family: "La Belle Aurore", cursive;
+           color: #ffffff;
+           font-size: 1.5rem;
+           line-height: 18px;
+           position: absolute;
+           left: -30px;
+           bottom: -20px;
+           margin-left: 20px;
+           opacity: 0.6;
+        }
+     }
+
+     &:nth-of-type(1) {
+        animation-delay: 1.1s;
+     }
+
+     &:nth-of-type(2) {
+        animation-delay: 1.2s;
+     }
+
+     &:nth-of-type(3) {
+        animation-delay: 1.3s;
+     }
+  }
+}
+
 
 @media screen and (max-width: 568px) {
     .tags {
@@ -18,19 +70,3 @@
     }
   }
 
-// @media screen and (max-width: 568px) {
-//     .container{
-//       transform: translateY(-1rem);
-//       padding: 0;
-//       margin: 0;
-//     }
-//     .container.about-page .text-zone {
-//         left: 65%;
-//     }
-//     .tags {
-//         display: none !important;
-//       }
-//       .nav-bar{
-//         height:129%;
-//       }
-// }
