--- src/components/Card/index.scss
@@ -1,131 +1,0 @@
-*
-{
-    font-family: 'Helvetica Neue', sans-serif;
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-.card{
-    position: relative;
-    width: 350px;
-    height: 190px;
-    background: fff;
-    border-radius: 20px;
-    box-shadow: 0 35px 80px rgba(0,0,0,0.15);
-    transition: 0.5s;
-}
-.card:hover{
-    height: 450px;
-    
-}
-.image-box{
-    position: absolute;
-    left: 50%;
-    top: -50px;
-    transform: translateX(-50%);
-    width: 150px;
-    height: 150px;
-    background: #fff;
-    border-radius: 20px;
-    box-shadow: 0 15px 50px rgba(0,0,0,0.35);
-    overflow: hidden;
-    transition: 0.5s;
-}
-
-.card:hover .image-box{
-    width: 250px;
-    height: 250px;
-}
-
-.image-box img{
-position: absolute;
-top: 0;
-left: 0;
-width: 100%;
-height: 100%;
-object-fit: cover;
-
-}
-
-.card .content{
-position: absolute;
-width: 100%;
-height: 100%;
-display: flex;
-justify-content: center;
-align-items: flex-end;
-overflow: hidden;
-}
-
-.card .content .details {
-    padding: 40px;
-    text-align: center;
-    width: 100%;
-    transition: 0.5s;
-    transform: translateY(150px);
-}
-
-.card:hover .content .details{
-    transform: translateY(0px);
-}
-
-.card .content .details h2 {
-    font-size: 1.25rem;
-    font-weight: 600;
-    stop-color: #555;
-    line-height: 1.2rem;
-}
-
-.card .content .details h2 span{
-    font-size: 0.75rem;
-    font-weight: 500;
-    opacity: 0.5;
-}
-
-.card .content .details .data{
-    display: flex;
-    justify-content: space-between;
-    margin: 20px 0;
-}
-
-.card .content .details .data h3{
-    font-size: 1rem;
-    color: #555;
-    line-height: 1.2rem;
-    font-weight: 600;
-
-}
-
-.card .content .details .data h3 span{
-font-size: 0.85rem;
-font-weight: 400;
-opacity: 0.5;
-}
-
-.card .content .details .actionBtn {
-    display: flex;
-    justify-content: space-between;
-   // gap: 20px;
-}
-
-.card .content .details .actionBtn button{
-    padding: 10px 30px;
-    border: none;
-    border-radius: 5px;
-    outline: none;
-    font-size: 1rem;
-    font-weight: 500;
-    background: ff5f95;
-    color: #fff;
-    cursor: pointer;
-}
-
-.card .content .details .actionBtn button:nth-child(2){
-    border: 1px solid #999;
-    color: #999;
-    background: #fff;
-
-}
-
-
