+++ src/components/Contact/index.js
@@ -109,7 +109,7 @@ const Contact = () => {
                </MapContainer>
             </div>
          </div>
-         <Loader type="pacman" />
+         <Loader type="line-scale-pulse-out-rapid" />
       </>
    )
 }
