+++ src/components/Contact/index.scss
@@ -134,7 +134,7 @@
     }
     .map-wrap{
       top: 5rem;
-      margin-left: -6.4rem;
+      margin-left: -4.4rem;
       width: 94%;
       float:none;
       margin-top: 19rem;
