+++ src/components/EmployeeList/index.scss
@@ -1,94 +1,3 @@
-/*
-
-*{
-	margin: 0;
-	padding: 0;
-	box-sizing: border-box;
-	font-family: 'Noto Sans JP', sans-serif;
-}
-  .cards_wrap{
-    margin-bottom: 10rem;
-    margin-top: 10rem;
-    padding: 6rem 1rem 2rem 7rem;    
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-    flex-wrap: wrap;
-  }
-  
-  .cards_wrap .card_item{
-    padding: 15px 25px;
-    width: 25%;
-  }
-  
-  .cards_wrap .card_inner{
-    background: #fff;
-    border-radius: 5px;
-    text-align: center;
-    padding: 35px 20px;
-    min-width: 225px;
-    min-height: 315px;
-    max-height: 370px;
-    width: 100%;
-  }
-  
-  .cards_wrap .card_item img{
-    width: 12rem;
-    height: 12rem;
-    margin-bottom: 5px;
-    transform: translateY(-8rem);
-;
-  }
-  
-  .cards_wrap .card_item .role_name{
-    color: #e36686;
-    font-weight: 900;
-    letter-spacing: 2px;
-    text-transform: uppercase;
-    font-size: 20px;
-    white-space: nowrap;
-      overflow: hidden;
-      text-overflow: ellipsis;
-  }
-  
-  .cards_wrap .card_item .real_name{
-    color: #b6c0c2;
-    font-size: 12px;
-    font-weight: 100;
-    margin: 5px 0 10px;
-  }
-  
-  .cards_wrap .card_item .film{
-    font-size: 14px;
-    line-height: 24px;
-    color: #7b8ca0;
-  }
-  
-  @media screen and (max-width: 1024px){
-    .cards_wrap .card_item{
-      width: 33%;
-    }
-  }
-  
-  @media screen and (max-width: 768px){
-    .cards_wrap .card_item{
-      width: 50%;
-    }
-    .wrapper .header{
-      font-size: 16px;
-      height: 60px;
-    }
-  }
-  
-  @media screen and (max-width: 568px){
-    .cards_wrap .card_item{
-      width: 100%;
-    }
-    .wrapper .header{
-      font-size: 14px;
-    }
-  }
-  */
 @import url('https://fonts.googleapis.com/css2?family=Baloo+Da+2&display=swap');
 @import url('https://fonts.googleapis.com/css2?family=Noto+Sans+JP:wght@100;400;900&display=swap');
 
@@ -139,7 +48,7 @@ body {
    border-radius: 20px;
 
    min-width: 225px;
-   min-height: 315px;
+   min-height: 380px;
    max-height: fit-content;
    width: 100%;
    box-shadow: 0 35px 80px rgba(0, 0, 0, 0.15);
@@ -217,7 +126,7 @@ body {
    }
 
    .cards_wrap {
-      transform: translateX(-248px);
+      transform: translateX(-203px);
       padding-top: 3rem;
       width: 100%;
       display: flex;
