+++ src/components/Home/index.js
@@ -1,9 +1,7 @@
 import { useEffect, useState } from "react";
-// import { Link } from "react-router-dom";
 import Loader from "react-loaders";
 import AnimatedLetters from "../AnimatedLetters";
-//import LogoTitle from '../../assets/images/logo-s.png'
-//import Logo from './Logo'
+
 import "./index.scss";
 
 const Home = () => {
@@ -42,7 +40,7 @@ const Home = () => {
         </div>
       </div>
 
-         <Loader type="ball-scale-ripple" />
+      <Loader type="line-scale-pulse-out-rapid" />
       </>
    );
 };
