+++ src/components/Home/index.scss
@@ -58,7 +58,6 @@
       margin-top: 4rem;
       font-weight: 400;
       font-size: 0.9rem;
-      // font-family: sans-serif;
       font-family: 'Baloo Da 2', cursive;
       letter-spacing: 3px;
       animation: fadeIn 1s 1.8s backwards;
