+++ src/components/Layout/index.js
@@ -1,4 +1,3 @@
-// import { useEffect } from 'react'
 import { useState } from 'react'
 import { Outlet } from 'react-router-dom'
 import Sidebar from '../Sidebar/'
