--- src/components/Navigation/MainHeader.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import './MainHeader.css';
-
-const MainHeader = props => {
-  return <header className="main-header">{props.children}</header>;
-};
-
-export default MainHeader;
