+++ src/components/Projects/index.scss
@@ -1,3 +1,68 @@
+.container.projects-page {
+  .text-zone {
+     position: relative;
+     left: 75%;
+     top: 1.7rem;
+     width: 100% !important;
+     vertical-align: middle;
+     display: table-cell;
+     max-height: 90%;
+
+     h1 {
+        font-size: 60px;
+        font-family: 'Leckerli One', cursive;
+        color: #9448A2;
+        font-weight: 400;
+        margin-top: 0;
+        position: relative;
+        left: -10%;
+        top: 90%;
+        transform: translateY(60%);
+        width: 100% !important;
+        vertical-align: middle;
+        display: table-cell;
+        max-height: 90%;
+
+        &:before {
+           content: "<h1>";
+           font-family: "La Belle Aurore", cursive;
+           color: #ffffff;
+           font-size: 1.5rem;
+           position: absolute;
+           margin-top: -10px;
+           left: -10px;
+           opacity: 0.6;
+           line-height: 18px;
+        }
+
+        &:after {
+           content: "<h1/>";
+           font-family: "La Belle Aurore", cursive;
+           color: #ffffff;
+           font-size: 1.5rem;
+           line-height: 18px;
+           position: absolute;
+           left: -30px;
+           bottom: -20px;
+           margin-left: 20px;
+           opacity: 0.6;
+        }
+     }
+
+     &:nth-of-type(1) {
+        animation-delay: 1.1s;
+     }
+
+     &:nth-of-type(2) {
+        animation-delay: 1.2s;
+     }
+
+     &:nth-of-type(3) {
+        animation-delay: 1.3s;
+     }
+  }
+}
+
 .portfolio-section{
     width: 100%;
     position: relative;
@@ -11,6 +76,7 @@
 @media screen and (max-width: 568px) {
     .portfolio-section{
       top:8rem;
-      width: 95%;
+      width: 80%;
+      transform: translateX(70px);
     }
 }
