+++ src/components/Sidebar/index.scss
@@ -280,14 +280,6 @@
       }
     }
   }
-
-  
-
-
-
-
-
-
   .main-navigation__menu-btn span {
     display: block;
     width: 3rem;
@@ -304,9 +296,6 @@
 
     }
   }
-  // .main-navigation__header-nav {
-  //   display: none;
-  // }
 
   @media (min-width: 768px) {
     .main-navigation__menu-btn {
