--- src/components/UIElements/Backdrop.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-
-import './Backdrop.css';
-
-const Backdrop = props => {
-  return ReactDOM.createPortal(
-    <div className="backdrop" onClick={props.onClick}></div>,
-    document.getElementById('backdrop-hook')
-  );
-};
-
-export default Backdrop;
