+++ src/components/portfolio/portfolio.scss
@@ -1,4 +1,3 @@
-//@import "../../global.scss";
 $mainColor: transparent;
 
 $width: 768px;
@@ -85,7 +84,7 @@ $width: 768px;
         width: 100%;
         height: 100%;
         object-fit: cover;
-        z-index: 1;
+        z-index: -1;
         border-radius: 10%;
       }
