--- src/data.js
@@ -1,37 +1,0 @@
-
-export const employees = [
-   {
-      id: 1,
-      imgURL: "https://www.information.dk/sites/information.dk/files/styles/440x/public/media/2010/03/09/chux.jpg?itok=wyYPkVR8",
-      name: "Henrik",
-      title: "Alders President",
-      description: 'Expert in online solutions, with a range from Java to React with integration to numerous database technologies. No project is too big or too small, Henrik got it all',
-      linkedInUrl: "http://lama.dev",
-   },
-   {
-      id: 2,
-      imgURL: "../../public/logo512.png",
-      name: "Rikke",
-      title: "PIGE",
-      description: "Er god til katte og enhjørninger",
-      linkedInUrl: "http://lama.dev",
-   },
-   {
-      id: 3,
-      imgURL: "../../public/logo512.png",
-      name: "Rune",
-      title: "Kuvøseguff",
-      description: "Er en haj til FIGMA og udregning af billån",
-      linkedInUrl: "http://lama.dev",
-   },
-   {
-      id: 4,
-      imgURL: "../../public/logo512.png",
-      name: "Lars",
-      title: "CEO",
-      description: "Tilpas bredskulderet til at bære hele firmaet",
-      linkedInUrl: "http://lama.dev",
-   },
-
-
-];
