+++ src/employeeData.js
@@ -1,0 +1,36 @@
+export const employees = [
+   {
+      id: 1,
+      imgURL: "https://imgur.com/dHciFSK.png",
+      name: "Henrik",
+      title: "Alders President",
+      description: 'Expert in online solutions, with a range from Java to React with integration to numerous database technologies. No project is too big or too small, Henrik got it all.',
+      linkedInUrl: "http://lama.dev",
+   },
+   {
+      id: 2,
+      imgURL: "https://i.imgur.com/fBkw5dO.png",
+      name: "Rikke",
+      title: "PIGE",
+      description: "A classic Jack of All Trades! With programming skills that have yet to exceed her Japanese skills, but high proficiency in colourful, childish designs, what could possibly go wrong?",
+      linkedInUrl: "http://lama.dev",
+   },
+   {
+      id: 3,
+      imgURL: "https://i.imgur.com/Bh4kX3P.png",
+      name: "Rune",
+      title: "Kuvøseguff",
+      description: "An exciting cocktail containing the physical manifestations of a ux god, a degree in Financial Management and a lowkey addiction to World of Warcraft.",
+      linkedInUrl: "http://lama.dev",
+   },
+   {
+      id: 4,
+      imgURL: "https://i.imgur.com/KP9eKLL.png",
+      name: "Lars",
+      title: "CEO",
+      description: "I am specialized in automating borring stuff and data visualization. No data is too big for me, I can handle handle it all.",
+      linkedInUrl: "http://lama.dev",
+   },
+
+
+];
