+++ src/index.css
@@ -1,18 +1,4 @@
 body {
-   /* margin: 0;
-   min-height: 100%;
-   font: 300 11px/1.4 'Indie Flower', sans-serif;
-   color: #444;
-   -webkit-font-smoothing: antialiased;
-   -moz-osx-font-smoothing: grayscale;
-   overflow: hidden;
-   display: block;
-   background-size: cover;
-   background-repeat: no-repeat;
-   background-position: center center;
-   background-color: rgb(50, 50, 50); */
-
-
    margin: 0;
    min-height: 100%;
    font: 300 11px/1.4 'Indie Flower', sans-serif;
@@ -21,6 +7,4 @@ body {
    -moz-osx-font-smoothing: grayscale;
    display: block;
    background-color: rgb(50, 50, 50);
-   /* overflow: hidden; */
-   /* overflow-y: scroll; */
 }
