+++ src/projectsData.js
@@ -43,38 +43,7 @@ export const allPortfolio = [
   },
 ];
 
-// export const javaPortfolio = [
-//   {
-//     id: 1,
-//     title: "Web Social Media App",
-//     img: "https://cdn.dribbble.com/users/5031392/screenshots/15467520/media/c36b3b15b25b1e190d081abdbbf947cf.png?compress=1&resize=1200x900",
-//   },
-//   {
-//     id: 2,
-//     title: "Web Rampa UI Design",
-//     img: "https://cdn.dribbble.com/users/3967258/screenshots/15463803/media/4fddb9a2caf3b3bd634060f706a91e73.png?compress=1&resize=1200x900",
-//   },
-//   {
-//     id: 3,
-//     title: "Web E-commerce Design",
-//     img: "https://cdn.dribbble.com/users/3637068/screenshots/15467010/media/69a3279182b75365013fe285733d1122.png?compress=1&resize=1200x900",
-//   },
-//   {
-//     id: 4,
-//     title: "Web Relax App",
-//     img: "https://cdn.dribbble.com/users/2264844/screenshots/15463998/media/d85e92a332c41094f68e1f716884f7ce.jpg?compress=1&resize=1200x900",
-//   },
-//   {
-//     id: 5,
-//     title: "Web Design",
-//     img: "https://cdn.dribbble.com/users/387658/screenshots/15458608/media/e92ec1dd33dd1170ae3dc9c5272f2693.jpg?compress=1&resize=1200x900",
-//   },
-//   {
-//     id: 6,
-//     title: "Web Banking App",
-//     img: "https://cdn.dribbble.com/users/6384483/screenshots/15468426/media/ce9479fa1f8dba3a4a49840d76e55e31.png?compress=1&resize=1200x900",
-//   },
-// ];
+
 
 export const frontendPortfolio = [
   {
